@@ -32,17 +32,23 @@
         "doctrine/annotations": "~1.0"
     },
     "require-dev": {
+        "symfony/browser-kit": "~2.3",
+        "symfony/console": "~2.0",
         "symfony/finder": "~2.0",
         "symfony/security": "~2.4",
-        "symfony/form": "~2.3",
+        "symfony/form": "2.5.*",
         "symfony/class-loader": "~2.1",
-        "symfony/validator": "~2.1"
+        "symfony/expression-language": "~2.4",
+        "symfony/process": "~2.0",
+        "symfony/validator": "~2.5",
+        "symfony/yaml": "~2.0"
     },
     "suggest": {
         "symfony/console": "For using the console commands",
         "symfony/finder": "For using the translation loader and cache warmer",
         "symfony/form": "For using forms",
         "symfony/validator": "For using validation",
+        "symfony/yaml": "For using the config:debug and yaml:lint commands",
         "doctrine/cache": "For using alternative cache drivers"
     },
     "autoload": {
