@@ -23,7 +23,7 @@
         "symfony/property-access": "~2.3"
     },
     "require-dev": {
-        "symfony/validator": "~2.2",
+        "symfony/validator": "~2.5",
         "symfony/http-foundation": "~2.2",
         "symfony/http-kernel": "~2.4",
         "symfony/security-csrf": "~2.4",
