@@ -32,18 +32,46 @@ class __TwigTemplate_0511db74b80fd4019a2ea49d4c6c9bb0769b5effa88888069881913a924
         echo "    ";
         $this->displayParentBlock("stylesheets", $context, $blocks);
         echo "
-    <link rel=\"stylesheet\" href=\"";
+    ";
         // line 4
-        echo twig_escape_filter($this->env, $this->env->getExtension('assets')->getAssetUrl("bundles/bloggerblog/css/blog.css"), "html", null, true);
-        echo "\" />
-";
+        if (isset($context['assetic']['debug']) && $context['assetic']['debug']) {
+            // asset "4ce5d7f_0"
+            $context["asset_url"] = isset($context['assetic']['use_controller']) && $context['assetic']['use_controller'] ? $this->env->getExtension('routing')->getPath("_assetic_4ce5d7f_0") : $this->env->getExtension('assets')->getAssetUrl("_controller/css/blogger_part_1_blog_1.css");
+            // line 8
+            echo "        <link rel=\"stylesheet\" href=\"";
+            echo twig_escape_filter($this->env, (isset($context["asset_url"]) ? $context["asset_url"] : $this->getContext($context, "asset_url")), "html", null, true);
+            echo "\" media=\"screen\" />
+    ";
+            // asset "4ce5d7f_1"
+            $context["asset_url"] = isset($context['assetic']['use_controller']) && $context['assetic']['use_controller'] ? $this->env->getExtension('routing')->getPath("_assetic_4ce5d7f_1") : $this->env->getExtension('assets')->getAssetUrl("_controller/css/blogger_part_1_screen_2.css");
+            echo "        <link rel=\"stylesheet\" href=\"";
+            echo twig_escape_filter($this->env, (isset($context["asset_url"]) ? $context["asset_url"] : $this->getContext($context, "asset_url")), "html", null, true);
+            echo "\" media=\"screen\" />
+    ";
+            // asset "4ce5d7f_2"
+            $context["asset_url"] = isset($context['assetic']['use_controller']) && $context['assetic']['use_controller'] ? $this->env->getExtension('routing')->getPath("_assetic_4ce5d7f_2") : $this->env->getExtension('assets')->getAssetUrl("_controller/css/blogger_part_1_sidebar_3.css");
+            echo "        <link rel=\"stylesheet\" href=\"";
+            echo twig_escape_filter($this->env, (isset($context["asset_url"]) ? $context["asset_url"] : $this->getContext($context, "asset_url")), "html", null, true);
+            echo "\" media=\"screen\" />
+    ";
+        } else {
+            // asset "4ce5d7f"
+            $context["asset_url"] = isset($context['assetic']['use_controller']) && $context['assetic']['use_controller'] ? $this->env->getExtension('routing')->getPath("_assetic_4ce5d7f") : $this->env->getExtension('assets')->getAssetUrl("_controller/css/blogger.css");
+            echo "        <link rel=\"stylesheet\" href=\"";
+            echo twig_escape_filter($this->env, (isset($context["asset_url"]) ? $context["asset_url"] : $this->getContext($context, "asset_url")), "html", null, true);
+            echo "\" media=\"screen\" />
+    ";
+        }
+        unset($context["asset_url"]);
     }
 
-    // line 6
+    // line 11
     public function block_sidebar($context, array $blocks = array())
     {
-        // line 7
-        echo "    Sidebar content 
+        // line 12
+        echo "    ";
+        echo $this->env->getExtension('http_kernel')->renderFragment($this->env->getExtension('http_kernel')->controller("BloggerBlogBundle:Page:sidebar"));
+        echo " 
 ";
     }
 
@@ -59,6 +87,6 @@ class __TwigTemplate_0511db74b80fd4019a2ea49d4c6c9bb0769b5effa88888069881913a924
 
     public function getDebugInfo()
     {
-        return array (  46 => 7,  43 => 6,  37 => 4,  32 => 3,  29 => 2,);
+        return array (  72 => 12,  69 => 11,  41 => 8,  37 => 4,  32 => 3,  29 => 2,);
     }
 }
