@@ -40,7 +40,7 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
         ";
         // line 10
         $this->displayBlock('stylesheets', $context, $blocks);
-        // line 15
+        // line 19
         echo "        <link rel=\"shortcut icon\" href=\"";
         echo twig_escape_filter($this->env, $this->env->getExtension('assets')->getAssetUrl("favicon.ico"), "html", null, true);
         echo "\" />
@@ -51,18 +51,18 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
             <header id=\"header\">
                 <div class=\"top\">
                     ";
-        // line 22
+        // line 26
         $this->displayBlock('navigation', $context, $blocks);
-        // line 31
+        // line 35
         echo "                </div>
 
                 <hgroup>
                     <h2>";
-        // line 34
+        // line 38
         $this->displayBlock('blog_title', $context, $blocks);
         echo "</h2>
                     <h3>";
-        // line 35
+        // line 39
         $this->displayBlock('blog_tagline', $context, $blocks);
         echo "</h3>
                 </hgroup>
@@ -70,29 +70,29 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
 
             <section class=\"main-col\">
                 ";
-        // line 40
+        // line 44
         $this->displayBlock('body', $context, $blocks);
-        // line 41
+        // line 45
         echo "            </section>
             <aside class=\"sidebar\">
                 ";
-        // line 43
+        // line 47
         $this->displayBlock('sidebar', $context, $blocks);
-        // line 44
+        // line 48
         echo "            </aside>
 
             <div id=\"footer\">
                 ";
-        // line 47
+        // line 51
         $this->displayBlock('footer', $context, $blocks);
-        // line 50
+        // line 54
         echo "            </div>
         </section>
 
         ";
-        // line 53
+        // line 57
         $this->displayBlock('javascripts', $context, $blocks);
-        // line 54
+        // line 64
         echo "    </body>
 </html>
 ";
@@ -110,29 +110,39 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
         // line 11
         echo "            <link href='http://fonts.googleapis.com/css?family=Irish+Grover' rel='stylesheet' type='text/css'>
             <link href='http://fonts.googleapis.com/css?family=La+Belle+Aurore' rel='stylesheet' type='text/css'>
-            <link href=\"";
+            ";
         // line 13
-        echo twig_escape_filter($this->env, $this->env->getExtension('assets')->getAssetUrl("bundles/bloggerblog/css/screen.css"), "html", null, true);
-        echo "\" type=\"text/css\" rel=\"stylesheet\" />
-        ";
+        if (isset($context['assetic']['debug']) && $context['assetic']['debug']) {
+        } else {
+            // asset "51c56cc"
+            $context["asset_url"] = isset($context['assetic']['use_controller']) && $context['assetic']['use_controller'] ? $this->env->getExtension('routing')->getPath("_assetic_51c56cc") : $this->env->getExtension('assets')->getAssetUrl("_controller/css/51c56cc.css");
+            // line 16
+            echo "        <link href=\"";
+            echo twig_escape_filter($this->env, (isset($context["asset_url"]) ? $context["asset_url"] : $this->getContext($context, "asset_url")), "html", null, true);
+            echo "\" rel=\"stylesheet\" media=\"screen\" />
+    ";
+        }
+        unset($context["asset_url"]);
+        // line 18
+        echo "        ";
     }
 
-    // line 22
+    // line 26
     public function block_navigation($context, array $blocks = array())
     {
-        // line 23
+        // line 27
         echo "                        <nav>
                             <ul class=\"navigation\">
                                 <li><a href=\"";
-        // line 25
+        // line 29
         echo $this->env->getExtension('routing')->getPath("BloggerBlogBundle_homepage");
         echo "\">Home</a></li>
                                 <li><a href=\"";
-        // line 26
+        // line 30
         echo $this->env->getExtension('routing')->getPath("BloggerBlogBundle_about");
         echo "\">About</a></li>
                                 <li><a href=\"";
-        // line 27
+        // line 31
         echo $this->env->getExtension('routing')->getPath("BloggerBlogBundle_contact");
         echo "\">Contact</a></li>
                             </ul>
@@ -140,7 +150,7 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
                     ";
     }
 
-    // line 34
+    // line 38
     public function block_blog_title($context, array $blocks = array())
     {
         echo "<a href=\"";
@@ -148,7 +158,7 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
         echo "\">symblog</a>";
     }
 
-    // line 35
+    // line 39
     public function block_blog_tagline($context, array $blocks = array())
     {
         echo "<a href=\"";
@@ -156,27 +166,42 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
         echo "\">creating a blog in Symfony2</a>";
     }
 
-    // line 40
+    // line 44
     public function block_body($context, array $blocks = array())
     {
     }
 
-    // line 43
+    // line 47
     public function block_sidebar($context, array $blocks = array())
     {
     }
 
-    // line 47
+    // line 51
     public function block_footer($context, array $blocks = array())
     {
-        // line 48
+        // line 52
         echo "                    Symfony2 blog tutorial - created by <a href=\"https://github.com/dsyph3r\">dsyph3r</a>
                 ";
     }
 
-    // line 53
+    // line 57
     public function block_javascripts($context, array $blocks = array())
     {
+        // line 58
+        echo "            ";
+        if (isset($context['assetic']['debug']) && $context['assetic']['debug']) {
+        } else {
+            // asset "3c62cfd"
+            $context["asset_url"] = isset($context['assetic']['use_controller']) && $context['assetic']['use_controller'] ? $this->env->getExtension('routing')->getPath("_assetic_3c62cfd") : $this->env->getExtension('assets')->getAssetUrl("_controller/js/3c62cfd.js");
+            // line 61
+            echo "             <script type=\"text/javascript\" src=\"";
+            echo twig_escape_filter($this->env, (isset($context["asset_url"]) ? $context["asset_url"] : $this->getContext($context, "asset_url")), "html", null, true);
+            echo "\"></script>
+            ";
+        }
+        unset($context["asset_url"]);
+        // line 63
+        echo "        ";
     }
 
     public function getTemplateName()
@@ -191,6 +216,6 @@ class __TwigTemplate_bcbb916285de0d842a96d4c744cf33e6d773e5f761cf3fba7498d3aced0
 
     public function getDebugInfo()
     {
-        return array (  178 => 53,  173 => 48,  170 => 47,  165 => 43,  160 => 40,  152 => 35,  144 => 34,  136 => 27,  132 => 26,  128 => 25,  124 => 23,  121 => 22,  115 => 13,  111 => 11,  108 => 10,  102 => 6,  96 => 54,  94 => 53,  89 => 50,  87 => 47,  82 => 44,  80 => 43,  76 => 41,  74 => 40,  66 => 35,  62 => 34,  57 => 31,  55 => 22,  44 => 15,  42 => 10,  35 => 6,  28 => 1,);
+        return array (  204 => 63,  197 => 61,  191 => 58,  188 => 57,  183 => 52,  180 => 51,  175 => 47,  170 => 44,  162 => 39,  154 => 38,  146 => 31,  142 => 30,  138 => 29,  134 => 27,  131 => 26,  127 => 18,  120 => 16,  115 => 13,  111 => 11,  108 => 10,  102 => 6,  96 => 64,  94 => 57,  89 => 54,  87 => 51,  82 => 48,  80 => 47,  76 => 45,  74 => 44,  66 => 39,  62 => 38,  57 => 35,  55 => 26,  44 => 19,  42 => 10,  35 => 6,  28 => 1,);
     }
 }
