@@ -24,17 +24,17 @@ class __TwigTemplate_21409e46f945f93d087125f1719a147f381772dad0331266387f53965f2
         foreach ($context['_seq'] as $context["_key"] => $context["log"]) {
             // line 3
             echo "        <li";
-            if (($this->getAttribute((isset($context["log"]) ? $context["log"] : null), "priority") >= 400)) {
+            if (($this->getAttribute($context["log"], "priority", array()) >= 400)) {
                 echo " class=\"error\"";
-            } elseif (($this->getAttribute((isset($context["log"]) ? $context["log"] : null), "priority") >= 300)) {
+            } elseif (($this->getAttribute($context["log"], "priority", array()) >= 300)) {
                 echo " class=\"warning\"";
             }
             echo ">
             ";
             // line 4
-            echo twig_escape_filter($this->env, $this->getAttribute((isset($context["log"]) ? $context["log"] : null), "priorityName"), "html", null, true);
+            echo twig_escape_filter($this->env, $this->getAttribute($context["log"], "priorityName", array()), "html", null, true);
             echo " - ";
-            echo twig_escape_filter($this->env, $this->getAttribute((isset($context["log"]) ? $context["log"] : null), "message"), "html", null, true);
+            echo twig_escape_filter($this->env, $this->getAttribute($context["log"], "message", array()), "html", null, true);
             echo "
         </li>
     ";
@@ -59,6 +59,6 @@ class __TwigTemplate_21409e46f945f93d087125f1719a147f381772dad0331266387f53965f2
 
     public function getDebugInfo()
     {
-        return array (  26 => 3,  98 => 40,  93 => 9,  88 => 6,  78 => 40,  46 => 7,  27 => 4,  40 => 8,  44 => 9,  35 => 4,  31 => 4,  43 => 8,  41 => 5,  28 => 3,  201 => 92,  196 => 90,  183 => 82,  171 => 73,  166 => 71,  163 => 70,  158 => 67,  156 => 66,  151 => 63,  142 => 59,  138 => 57,  136 => 56,  123 => 47,  121 => 46,  117 => 44,  115 => 43,  105 => 40,  101 => 39,  91 => 31,  69 => 25,  66 => 24,  62 => 23,  49 => 19,  24 => 2,  32 => 6,  25 => 3,  22 => 2,  19 => 1,  209 => 82,  203 => 78,  199 => 91,  193 => 73,  189 => 71,  187 => 84,  182 => 68,  176 => 64,  173 => 74,  168 => 72,  164 => 60,  162 => 59,  154 => 54,  149 => 51,  147 => 50,  144 => 49,  141 => 48,  133 => 55,  130 => 41,  125 => 38,  122 => 37,  116 => 36,  112 => 42,  109 => 34,  106 => 33,  103 => 32,  99 => 30,  95 => 28,  92 => 27,  86 => 28,  82 => 22,  80 => 41,  73 => 19,  64 => 15,  60 => 13,  57 => 12,  54 => 11,  51 => 9,  48 => 9,  45 => 8,  42 => 6,  39 => 16,  36 => 7,  33 => 4,  30 => 3,);
+        return array (  46 => 7,  35 => 4,  26 => 3,  22 => 2,  19 => 1,);
     }
 }
