@@ -28,8 +28,80 @@ class __TwigTemplate_5ce91bc91ce2856c19fc34e2b10717d11a89d15eafb7d52aa13cd9cfe73
     public function block_body($context, array $blocks = array())
     {
         // line 4
-        echo "    Blog homepage
-";
+        echo "    ";
+        $context['_parent'] = (array) $context;
+        $context['_seq'] = twig_ensure_traversable((isset($context["blogs"]) ? $context["blogs"] : null));
+        $context['_iterated'] = false;
+        foreach ($context['_seq'] as $context["_key"] => $context["blog"]) {
+            // line 5
+            echo "        <article class=\"blog\">
+            <div class=\"date\">
+                <time datetime=\"";
+            // line 7
+            echo twig_escape_filter($this->env, twig_date_format_filter($this->env, $this->getAttribute($context["blog"], "createdAt", array()), "c"), "html", null, true);
+            echo "\">";
+            echo twig_escape_filter($this->env, twig_date_format_filter($this->env, $this->getAttribute($context["blog"], "createdAt", array()), "l, F j, Y"), "html", null, true);
+            echo "</time>
+            </div>
+            <header>
+                <h2><a href=\"";
+            // line 10
+            echo twig_escape_filter($this->env, $this->env->getExtension('routing')->getPath("BloggerBlogBundle_blog_show", array("id" => $this->getAttribute($context["blog"], "id", array()), "slug" => $this->getAttribute($context["blog"], "slug", array()))), "html", null, true);
+            echo "\">";
+            echo twig_escape_filter($this->env, $this->getAttribute($context["blog"], "title", array()), "html", null, true);
+            echo "</a></h2>
+            </header>
+            
+            <img src=\"";
+            // line 13
+            echo twig_escape_filter($this->env, $this->env->getExtension('assets')->getAssetUrl(twig_join_filter(array(0 => "images/", 1 => $this->getAttribute($context["blog"], "image", array())))), "html", null, true);
+            echo "\" />
+            <div class=\"snippet\">
+                <p>";
+            // line 15
+            echo twig_escape_filter($this->env, $this->getAttribute($context["blog"], "blog", array(0 => 500), "method"), "html", null, true);
+            echo "</p>
+                <p class=\"continue\">
+                    <a href=\"";
+            // line 17
+            echo twig_escape_filter($this->env, $this->env->getExtension('routing')->getPath("BloggerBlogBundle_blog_show", array("id" => $this->getAttribute($context["blog"], "id", array()), "slug" => $this->getAttribute($context["blog"], "slug", array()))), "html", null, true);
+            echo "\">Continue reading...</a>
+                </p>
+            </div>
+                
+                
+            <footer class=\"meta\">
+                <p>Comments: <a href=\"";
+            // line 23
+            echo twig_escape_filter($this->env, $this->env->getExtension('routing')->getPath("BloggerBlogBundle_blog_show", array("id" => $this->getAttribute($context["blog"], "id", array()), "slug" => $this->getAttribute($context["blog"], "slug", array()))), "html", null, true);
+            echo "#comments\">- ";
+            echo twig_escape_filter($this->env, twig_length_filter($this->env, $this->getAttribute($context["blog"], "comments", array())), "html", null, true);
+            echo "</a> </p>
+                <p>Posted by <span class=\"highlight\">";
+            // line 24
+            echo twig_escape_filter($this->env, $this->getAttribute($context["blog"], "author", array()), "html", null, true);
+            echo "</span> at ";
+            echo twig_escape_filter($this->env, twig_date_format_filter($this->env, $this->getAttribute($context["blog"], "createdAt", array()), "h:iA"), "html", null, true);
+            echo "</p>
+                <p>Tags: <span class=\"highlight\">";
+            // line 25
+            echo twig_escape_filter($this->env, $this->getAttribute($context["blog"], "tags", array()), "html", null, true);
+            echo "</span></p>
+            </footer>
+            
+            
+        </article>
+            ";
+            $context['_iterated'] = true;
+        }
+        if (!$context['_iterated']) {
+            // line 31
+            echo "        <p>There are no blog entries for symblog</p>
+    ";
+        }
+        $_parent = $context['_parent'];
+        unset($context['_seq'], $context['_iterated'], $context['_key'], $context['blog'], $context['_parent'], $context['loop']);
+        $context = array_intersect_key($context, $_parent) + $_parent;
     }
 
     public function getTemplateName()
@@ -44,6 +116,6 @@ class __TwigTemplate_5ce91bc91ce2856c19fc34e2b10717d11a89d15eafb7d52aa13cd9cfe73
 
     public function getDebugInfo()
     {
-        return array (  65 => 17,  52 => 13,  178 => 53,  170 => 47,  165 => 43,  160 => 40,  152 => 35,  124 => 23,  96 => 54,  76 => 41,  74 => 20,  480 => 162,  474 => 161,  469 => 158,  461 => 155,  457 => 153,  453 => 151,  444 => 149,  440 => 148,  437 => 147,  435 => 146,  430 => 144,  427 => 143,  423 => 142,  413 => 134,  409 => 132,  407 => 131,  402 => 130,  398 => 129,  393 => 126,  387 => 122,  384 => 121,  381 => 120,  379 => 119,  374 => 116,  368 => 112,  365 => 111,  362 => 110,  360 => 109,  355 => 106,  341 => 105,  337 => 103,  322 => 101,  314 => 99,  312 => 98,  309 => 97,  305 => 95,  298 => 91,  294 => 90,  285 => 89,  283 => 88,  278 => 86,  268 => 85,  264 => 84,  258 => 81,  252 => 80,  247 => 78,  241 => 77,  229 => 73,  220 => 70,  214 => 69,  177 => 65,  169 => 60,  140 => 55,  132 => 26,  128 => 25,  111 => 11,  107 => 36,  61 => 16,  273 => 96,  269 => 94,  254 => 92,  246 => 90,  243 => 88,  240 => 86,  238 => 85,  235 => 74,  230 => 82,  227 => 81,  224 => 71,  221 => 77,  219 => 76,  217 => 75,  208 => 68,  204 => 72,  179 => 69,  159 => 61,  143 => 56,  135 => 53,  131 => 52,  119 => 42,  108 => 10,  102 => 6,  71 => 19,  67 => 15,  63 => 15,  59 => 14,  47 => 9,  38 => 6,  94 => 53,  89 => 50,  85 => 25,  79 => 18,  75 => 17,  68 => 14,  56 => 9,  50 => 10,  29 => 3,  87 => 47,  72 => 16,  55 => 22,  21 => 2,  26 => 6,  98 => 31,  93 => 28,  88 => 6,  78 => 21,  46 => 12,  27 => 4,  40 => 8,  44 => 15,  35 => 5,  31 => 4,  43 => 8,  41 => 7,  28 => 3,  201 => 92,  196 => 90,  183 => 82,  171 => 61,  166 => 71,  163 => 62,  158 => 67,  156 => 66,  151 => 63,  142 => 59,  138 => 54,  136 => 27,  123 => 47,  121 => 22,  117 => 44,  115 => 13,  105 => 40,  101 => 32,  91 => 27,  69 => 18,  66 => 35,  62 => 34,  49 => 19,  24 => 4,  32 => 4,  25 => 3,  22 => 2,  19 => 1,  209 => 82,  203 => 78,  199 => 67,  193 => 73,  189 => 71,  187 => 84,  182 => 66,  176 => 64,  173 => 48,  168 => 72,  164 => 59,  162 => 59,  154 => 58,  149 => 51,  147 => 58,  144 => 34,  141 => 48,  133 => 55,  130 => 41,  125 => 44,  122 => 43,  116 => 41,  112 => 42,  109 => 34,  106 => 33,  103 => 32,  99 => 31,  95 => 28,  92 => 21,  86 => 28,  82 => 44,  80 => 43,  73 => 19,  64 => 17,  60 => 13,  57 => 15,  54 => 10,  51 => 14,  48 => 13,  45 => 8,  42 => 10,  39 => 9,  36 => 5,  33 => 4,  30 => 7,);
+        return array (  99 => 31,  88 => 25,  82 => 24,  76 => 23,  67 => 17,  62 => 15,  57 => 13,  49 => 10,  41 => 7,  37 => 5,  31 => 4,  28 => 3,);
     }
 }
