@@ -11,6 +11,7 @@ namespace Blogger\BlogBundle\Controller;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Blogger\BlogBundle\Entity\Comment;
 use Blogger\BlogBundle\Form\CommentType;
+use Symfony\Component\HttpFoundation\Request;
 
 /**
  * Description of CommentController
@@ -19,52 +20,59 @@ use Blogger\BlogBundle\Form\CommentType;
  */
 class CommentController extends Controller{
     
-    public function newAction($blog_id){
-        $blog= $this->getBlog($blog_id);
-        
+    public function newAction($blog_id)
+    {
+        $blog = $this->getBlog($blog_id);
+
         $comment = new Comment();
         $comment->setBlog($blog);
-        $form =$this->createForm(new CommentType(), $comment);
-        
-        return $this->render('BoggerBlogBundle:Comment:form.html.twig', array(
-            'comment'=>$comment,
-            'from'=>$form->createView()
+        $form   = $this->createForm(new CommentType(), $comment);
+
+        return $this->render('BloggerBlogBundle:Comment:form.html.twig', array(
+            'comment' => $comment,
+            'form'   => $form->createView()
         ));
     }
-    
-    public function createAction(Request $request, $blog_id){
+
+    public function createAction(Request $request, $blog_id)
+    {
         $blog = $this->getBlog($blog_id);
-        $comment = new Comment();
+
+        $comment  = new Comment();
         $comment->setBlog($blog);
-        $form = $this->createForm(new CommentType(),$comment);
         
-        if($form->isValid()){
-            $form->handleRequest($request);
-            
-            return $this->redirect($this->generateUrl('BloggerBlogBundle_blog_show',
-                    array('id'=>$comment->getBlog()->getId(),
-                            '#comment-' . $comment->getId()
-                        )));
+        $form    = $this->createForm(new CommentType(), $comment);
+        $form->handleRequest($request);
+
+        if ($form->isValid()) {
+            $em= $this->getDoctrine()->getManager();
+            $em->persist($comment);
+            $em->flush();
+
+            return $this->redirect($this->generateUrl('BloggerBlogBundle_blog_show', array(
+            'id'    => $comment->getBlog()->getId(),
+            'slug'  => $comment->getBlog()->getSlug())) .
+            '#comment-' . $comment->getId()
+        );
         }
-        
-        return $this->render('BloggerBlogBundle:Comment:create.html.twig',
-                array(
-                    'comment'=>$comment,
-                    'form'=>$form->createView()
-                ));
-        
-        
+
+        return $this->render('BloggerBlogBundle:Comment:create.html.twig', array(
+            'comment' => $comment,
+            'form'    => $form->createView()
+        ));
     }
-    
-    protected function getBlog($blog_id){
-        $em= $this->getDoctrine()->getManager();
-        
+
+    protected function getBlog($blog_id)
+    {
+        $em = $this->getDoctrine()
+                    ->getEntityManager();
+
         $blog = $em->getRepository('BloggerBlogBundle:Blog')->find($blog_id);
-        
-        if(!$blog_id){
-            throw $this->createNotFoundException('Unable to find Blog post');
+
+        if (!$blog) {
+            throw $this->createNotFoundException('Unable to find Blog post.');
         }
-        
+
         return $blog;
     }
 }
